requirements.txt
Django==3.1.5
djangorestframework==3.12.2
requests==2.25.1
fake-useragent==0.1.11
celery==5.0.5
redis==3.5.3
uWSGI==2.0.19.1